Add resume behavior for audio cutting and transcription
Running the whole process on a 1 hour file takes about 2 hours on my laptop. This makes the script
a bit more resilient, provided that speaker diarization has completed.

Speaker diarization can't guarantee the same speaker ids on a second run, so if the script crashes during
the diarization phase then it will have to start over.